refactor types and components structure
